Splitting up delta_generator, part 3! (#1805)
* Extract _iframe and _html

* Extract audio and video

* _pb2 to Proto, part trois

* Fix typo while extracting video

* Extract checkbox, multiselect, select, radio

* Extract time_input, date_input, text_input, and text_area

* Extract empty and progress

* Extract number_input

* Extract color_picker and file_uploader

* Extract out slider

* Fix mypy issues

* Extract out image and pyplot

I think we're done!

* Remove legacy @_with_element and _enqueue_new_element_delta

* Only import matplotlib once st.pyplot is actually called

* Rename beta_ namespaces in the actual mixin

* Shorten references of _enqueue_new_element_delta => _enqueue
