+++ lib/streamlit/__init__.py
@@ -100,7 +100,6 @@ from streamlit import env_util as _env_util
 from streamlit import source_util as _source_util
 from streamlit import string_util as _string_util
 from streamlit import type_util as _type_util
-from streamlit.commands.page_config import set_page_config as _set_page_config
 from streamlit.delta_generator import DeltaGenerator as _DeltaGenerator
 from streamlit.report_thread import add_report_ctx as _add_report_ctx
 from streamlit.report_thread import get_report_ctx as _get_report_ctx
@@ -187,7 +186,7 @@ beta_color_picker = _main.beta_color_picker  # noqa: E221
 # Config
 
 get_option = _config.get_option
-beta_set_page_config = _set_page_config
+from streamlit.commands.page_config import beta_set_page_config
 
 
 def set_option(key, value):
