+++ lib/streamlit/commands/page_config.py
@@ -19,7 +19,7 @@ from streamlit.elements import image_proto
 from streamlit.errors import StreamlitAPIException
 
 
-def set_page_config(
+def beta_set_page_config(
     page_title=None, page_icon=None, layout="centered", initial_sidebar_state="auto"
 ):
     """
