+++ lib/streamlit/elements/image_proto.py
@@ -22,7 +22,9 @@ import numpy as np
 
 from PIL import Image, ImageFile
 
-from streamlit.errors import StreamlitAPIException
+from streamlit import config
+from streamlit.proto.Image_pb2 import ImageList as ImageListProto
+from streamlit.errors import StreamlitAPIException, StreamlitDeprecationWarning
 from streamlit.logger import get_logger
 from urllib.parse import quote
 from urllib.parse import urlparse
@@ -36,6 +38,119 @@ LOGGER = get_logger(__name__)
 # 730 is the max width of element-container in the frontend, and 2x is for high
 # DPI.
 MAXIMUM_CONTENT_WIDTH = 2 * 730
+
+
+class ImageMixin:
+    def image(
+        dg,
+        image,
+        caption=None,
+        width=None,
+        use_column_width=False,
+        clamp=False,
+        channels="RGB",
+        output_format="auto",
+        **kwargs,
+    ):
+        """Display an image or list of images.
+
+        Parameters
+        ----------
+        image : numpy.ndarray, [numpy.ndarray], BytesIO, str, or [str]
+            Monochrome image of shape (w,h) or (w,h,1)
+            OR a color image of shape (w,h,3)
+            OR an RGBA image of shape (w,h,4)
+            OR a URL to fetch the image from
+            OR an SVG XML string like `<svg xmlns=...</svg>`
+            OR a list of one of the above, to display multiple images.
+        caption : str or list of str
+            Image caption. If displaying multiple images, caption should be a
+            list of captions (one for each image).
+        width : int or None
+            Image width. None means use the image width.
+            Should be set for SVG images, as they have no default image width.
+        use_column_width : bool
+            If True, set the image width to the column width. This takes
+            precedence over the `width` parameter.
+        clamp : bool
+            Clamp image pixel values to a valid range ([0-255] per channel).
+            This is only meaningful for byte array images; the parameter is
+            ignored for image URLs. If this is not set, and an image has an
+            out-of-range value, an error will be thrown.
+        channels : 'RGB' or 'BGR'
+            If image is an nd.array, this parameter denotes the format used to
+            represent color information. Defaults to 'RGB', meaning
+            `image[:, :, 0]` is the red channel, `image[:, :, 1]` is green, and
+            `image[:, :, 2]` is blue. For images coming from libraries like
+            OpenCV you should set this to 'BGR', instead.
+        output_format : 'JPEG', 'PNG', or 'auto'
+            This parameter specifies the format to use when transferring the
+            image data. Photos should use the JPEG format for lossy compression
+            while diagrams should use the PNG format for lossless compression.
+            Defaults to 'auto' which identifies the compression type based
+            on the type and format of the image argument.
+
+        Example
+        -------
+        >>> from PIL import Image
+        >>> image = Image.open('sunrise.jpg')
+        >>>
+        >>> st.image(image, caption='Sunrise by the mountains',
+        ...          use_column_width=True)
+
+        .. output::
+           https://share.streamlit.io/0.61.0-yRE1/index.html?id=Sn228UQxBfKoE5C7A7Y2Qk
+           height: 630px
+
+        """
+
+        format = kwargs.get("format")
+        if format != None:
+            # override output compression type if specified
+            output_format = format
+
+            if config.get_option("deprecation.showImageFormat"):
+                dg.exception(ImageFormatWarning(format))  # type: ignore
+
+        if use_column_width:
+            width = -2
+        elif width is None:
+            width = -1
+        elif width <= 0:
+            raise StreamlitAPIException("Image width must be positive.")
+
+        image_list_proto = ImageListProto()
+        marshall_images(
+            dg._get_coordinates(),  # type: ignore
+            image,
+            caption,
+            width,
+            image_list_proto,
+            clamp,
+            channels,
+            output_format,
+        )
+        return dg._enqueue("imgs", image_list_proto)  # type: ignore
+
+
+class ImageFormatWarning(StreamlitDeprecationWarning):
+    def __init__(self, format):
+        self.format = format
+
+        super(ImageFormatWarning, self).__init__(
+            msg=self._get_message(), config_option="deprecation.showImageFormat"
+        )
+
+    def _get_message(self):
+        return f"""
+The `format` parameter for `st.image` has been deprecated and will be removed
+or repurposed in the future. We recommend changing to the new `output_format`
+parameter to future-proof your code. For the parameter,
+`format="{self.format}"`, please use `output_format="{self.format}"` instead.
+
+See [https://github.com/streamlit/streamlit/issues/1137](https://github.com/streamlit/streamlit/issues/1137)
+for more information.
+            """
 
 
 def _image_has_alpha_channel(image):
