+++ lib/streamlit/elements/text_widgets.py
@@ -1,0 +1,112 @@
+from streamlit.proto.TextInput_pb2 import TextInput as TextInputProto
+from streamlit.proto.TextArea_pb2 import TextArea as TextAreaProto
+from streamlit.errors import StreamlitAPIException
+from .utils import _get_widget_ui_value
+
+
+class TextWidgetsMixin:
+    def text_input(dg, label, value="", max_chars=None, key=None, type="default"):
+        """Display a single-line text input widget.
+
+        Parameters
+        ----------
+        label : str
+            A short label explaining to the user what this input is for.
+        value : any
+            The text value of this widget when it first renders. This will be
+            cast to str internally.
+        max_chars : int or None
+            Max number of characters allowed in text input.
+        key : str
+            An optional string to use as the unique key for the widget.
+            If this is omitted, a key will be generated for the widget
+            based on its content. Multiple widgets of the same type may
+            not share the same key.
+        type : str
+            The type of the text input. This can be either "default" (for
+            a regular text input), or "password" (for a text input that
+            masks the user's typed value). Defaults to "default".
+
+        Returns
+        -------
+        str
+            The current value of the text input widget.
+
+        Example
+        -------
+        >>> title = st.text_input('Movie title', 'Life of Brian')
+        >>> st.write('The current movie title is', title)
+
+        """
+        text_input_proto = TextInputProto()
+        text_input_proto.label = label
+        text_input_proto.default = str(value)
+
+        if max_chars is not None:
+            text_input_proto.max_chars = max_chars
+
+        if type == "default":
+            text_input_proto.type = TextInputProto.DEFAULT
+        elif type == "password":
+            text_input_proto.type = TextInputProto.PASSWORD
+        else:
+            raise StreamlitAPIException(
+                "'%s' is not a valid text_input type. Valid types are 'default' and 'password'."
+                % type
+            )
+
+        ui_value = _get_widget_ui_value("text_input", text_input_proto, user_key=key)
+        current_value = ui_value if ui_value is not None else value
+        return dg._enqueue("text_input", text_input_proto, str(current_value))  # type: ignore
+
+    def text_area(dg, label, value="", height=None, max_chars=None, key=None):
+        """Display a multi-line text input widget.
+
+        Parameters
+        ----------
+        label : str
+            A short label explaining to the user what this input is for.
+        value : any
+            The text value of this widget when it first renders. This will be
+            cast to str internally.
+        height : int or None
+            Desired height of the UI element expressed in pixels. If None, a
+            default height is used.
+        max_chars : int or None
+            Maximum number of characters allowed in text area.
+        key : str
+            An optional string to use as the unique key for the widget.
+            If this is omitted, a key will be generated for the widget
+            based on its content. Multiple widgets of the same type may
+            not share the same key.
+
+        Returns
+        -------
+        str
+            The current value of the text input widget.
+
+        Example
+        -------
+        >>> txt = st.text_area('Text to analyze', '''
+        ...     It was the best of times, it was the worst of times, it was
+        ...     the age of wisdom, it was the age of foolishness, it was
+        ...     the epoch of belief, it was the epoch of incredulity, it
+        ...     was the season of Light, it was the season of Darkness, it
+        ...     was the spring of hope, it was the winter of despair, (...)
+        ...     ''')
+        >>> st.write('Sentiment:', run_sentiment_analysis(txt))
+
+        """
+        text_area_proto = TextAreaProto()
+        text_area_proto.label = label
+        text_area_proto.default = str(value)
+
+        if height is not None:
+            text_area_proto.height = height
+
+        if max_chars is not None:
+            text_area_proto.max_chars = max_chars
+
+        ui_value = _get_widget_ui_value("text_area", text_area_proto, user_key=key)
+        current_value = ui_value if ui_value is not None else value
+        return dg._enqueue("text_area", text_area_proto, str(current_value))  # type: ignore
